apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "auto-trade-api"
version = "0.1.0"
description = "Phase 2 backend foundation for auto-trade"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110.0,<1.0.0",
  "uvicorn[standard]>=0.29.0,<1.0.0",
  "pydantic-settings>=2.2.0,<3.0.0",
  "sqlalchemy[asyncio]>=2.0.25,<3.0.0",
  "asyncpg>=0.29.0,<1.0.0",
  "redis>=5.0.0,<6.0.0",
  "argon2-cffi>=23.1.0,<26.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0,<9.0.0",
  "pytest-asyncio>=0.23.0,<1.0.0",
  "httpx>=0.27.0,<1.0.0"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
